pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = "REST API for user management"
authors = ["Vladyslav Timofieiev <vladislav.tmf@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.1.0"
python-dotenv = "^1.0.1"
flask-marshmallow = "^1.3.0"
marshmallow-sqlalchemy = "^1.4.1"
flask-bcrypt = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
